mcp/pyproject.toml
[project]
name = "mcp-server-ask-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.2",
    "pyside6>=6.9.1",
    "pywebview[qt]>=5.4",
    "ttkthemes>=3.2.2",
]
